+++ packages/ui/src/components/DIalog/dialog.tsx
@@ -1,0 +1,250 @@
+import React from 'react';
+import * as DialogPrimitive from '@radix-ui/react-dialog';
+import { XIcon } from 'lucide-react';
+
+import { cn } from '@/lib/utils';
+import Button from '@/components/Button/Button';
+
+export interface DialogProps {
+  open: boolean;
+  onOpenChange: (open: boolean) => void;
+  title?: string;
+  children: React.ReactNode;
+  onConfirm?: () => void;
+  onCancel?: () => void;
+  confirmText?: string;
+  cancelText?: string;
+  buttonLayout?: 'horizontal' | 'vertical';
+  hideActions?: boolean;
+  closeOnBackdropClick?: boolean;
+  closeOnEscape?: boolean;
+  ariaLabel?: string;
+  className?: string;
+  showCloseButton?: boolean;
+}
+
+//내부 컴포넌트들
+function DialogRoot({ ...props }: React.ComponentProps<typeof DialogPrimitive.Root>) {
+  return <DialogPrimitive.Root {...props} />;
+}
+
+function DialogTrigger({ ...props }: React.ComponentProps<typeof DialogPrimitive.Trigger>) {
+  return <DialogPrimitive.Trigger {...props} />;
+}
+
+function DialogPortal({ ...props }: React.ComponentProps<typeof DialogPrimitive.Portal>) {
+  return <DialogPrimitive.Portal {...props} />;
+}
+
+function DialogClose({ ...props }: React.ComponentProps<typeof DialogPrimitive.Close>) {
+  return <DialogPrimitive.Close {...props} />;
+}
+
+function DialogOverlay({
+  className,
+  ...props
+}: React.ComponentProps<typeof DialogPrimitive.Overlay>) {
+  return (
+    <DialogPrimitive.Overlay
+      className={cn(
+        'data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 fixed inset-0 z-50 bg-black/50',
+        className
+      )}
+      {...props}
+    />
+  );
+}
+
+function DialogContent({
+  className,
+  children,
+  showCloseButton = true,
+  ...props
+}: React.ComponentProps<typeof DialogPrimitive.Content> & {
+  showCloseButton?: boolean;
+}) {
+  return (
+    <DialogPortal>
+      <DialogOverlay />
+      <DialogPrimitive.Content
+        className={cn(
+          'fixed left-[50%] top-[50%] z-50 grid w-full max-w-lg translate-x-[-50%] translate-y-[-50%] gap-4 border bg-white p-6 shadow-lg duration-200',
+          'data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0',
+          'data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95 sm:rounded-lg',
+          'max-w-[calc(100%-2rem)]',
+          className
+        )}
+        {...props}
+      >
+        {children}
+        {/* DialogContent에서는 닫기 버튼 제거 - Dialog 컴포넌트에서만 관리 */}
+      </DialogPrimitive.Content>
+    </DialogPortal>
+  );
+}
+
+function DialogHeader({ className, ...props }: React.ComponentProps<'div'>) {
+  return (
+    <div
+      className={cn('flex flex-col space-y-1.5 text-center sm:text-left', className)}
+      {...props}
+    />
+  );
+}
+
+function DialogFooter({ className, ...props }: React.ComponentProps<'div'>) {
+  return (
+    <div
+      className={cn('flex flex-col-reverse sm:flex-row sm:justify-end sm:space-x-2', className)}
+      {...props}
+    />
+  );
+}
+
+function DialogTitle({ className, ...props }: React.ComponentProps<typeof DialogPrimitive.Title>) {
+  return (
+    <DialogPrimitive.Title
+      className={cn('text-lg font-semibold leading-none tracking-tight', className)}
+      {...props}
+    />
+  );
+}
+
+function DialogDescription({
+  className,
+  ...props
+}: React.ComponentProps<typeof DialogPrimitive.Description>) {
+  return (
+    <DialogPrimitive.Description className={cn('text-sm text-gray-500', className)} {...props} />
+  );
+}
+
+const Dialog: React.FC<DialogProps> = ({
+  open = false,
+  onOpenChange,
+  title,
+  children,
+  onConfirm,
+  onCancel,
+  confirmText = '확인',
+  cancelText = '취소',
+  buttonLayout = 'horizontal',
+  hideActions = false,
+  closeOnBackdropClick = true,
+  closeOnEscape = true,
+  ariaLabel,
+  className,
+  showCloseButton = true,
+}) => {
+  const handleConfirm = () => {
+    if (onConfirm) {
+      onConfirm();
+    } else {
+      onOpenChange(false);
+    }
+  };
+
+  const handleCancel = () => {
+    if (onCancel) {
+      onCancel();
+    } else {
+      onOpenChange(false);
+    }
+  };
+
+  return (
+    <DialogPrimitive.Root open={open} onOpenChange={onOpenChange}>
+      <DialogPrimitive.Portal>
+        <DialogPrimitive.Overlay className="data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 fixed inset-0 z-50 bg-black/50" />
+        <DialogPrimitive.Content
+          className={cn(
+            'fixed left-[50%] top-[50%] z-50 grid translate-x-[-50%] translate-y-[-50%] gap-4 border bg-white p-6 shadow-lg duration-200',
+            'data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0',
+            'data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95 sm:rounded-lg',
+            'max-w-[calc(100%-2rem)]',
+            className
+          )}
+          aria-label={ariaLabel || title}
+          onInteractOutside={(event) => {
+            if (!closeOnBackdropClick) {
+              event.preventDefault();
+            }
+          }}
+          onEscapeKeyDown={(event) => {
+            if (!closeOnEscape) {
+              event.preventDefault();
+            }
+          }}
+        >
+          {/* 닫기 버튼 - 한 곳에서만 관리 */}
+          {showCloseButton && (
+            <DialogPrimitive.Close className="absolute right-4 top-4 z-10 rounded-sm opacity-70 ring-offset-white transition-opacity hover:opacity-100 focus:outline-none focus:ring-2 focus:ring-gray-500 focus:ring-offset-2 disabled:pointer-events-none">
+              <XIcon className="h-4 w-4" />
+              <span className="sr-only">Close</span>
+            </DialogPrimitive.Close>
+          )}
+
+          {/* 제목 - 닫기 버튼과 겹침 방지하면서 중앙 정렬 */}
+          {title && (
+            <DialogPrimitive.Title
+              className={cn(
+                'mb-4 whitespace-pre-line text-center text-lg font-semibold leading-none tracking-tight',
+                showCloseButton ? 'px-8 pt-8' : 'pt-2' // 🔧 양쪽 패딩으로 진짜 중앙 정렬
+              )}
+            >
+              {title}
+            </DialogPrimitive.Title>
+          )}
+
+          {/* 본문 */}
+          <div className="mb-6 text-center">{children}</div>
+
+          {/* 액션 버튼 */}
+          {!hideActions && (onConfirm || onCancel) && (
+            <div
+              className={cn(
+                'flex justify-center gap-3',
+                buttonLayout === 'vertical' ? 'flex-col' : 'flex-row'
+              )}
+            >
+              {onCancel && (
+                <Button
+                  variant="secondary"
+                  onClick={handleCancel}
+                  className="flex-1"
+                  aria-label={`${cancelText} 버튼`}
+                >
+                  {cancelText}
+                </Button>
+              )}
+              {onConfirm && (
+                <Button
+                  variant="primary"
+                  onClick={handleConfirm}
+                  className="flex-1"
+                  aria-label={`${confirmText} 버튼`}
+                >
+                  {confirmText}
+                </Button>
+              )}
+            </div>
+          )}
+        </DialogPrimitive.Content>
+      </DialogPrimitive.Portal>
+    </DialogPrimitive.Root>
+  );
+};
+
+export {
+  Dialog,
+  DialogRoot,
+  DialogClose,
+  DialogContent,
+  DialogDescription,
+  DialogFooter,
+  DialogHeader,
+  DialogOverlay,
+  DialogPortal,
+  DialogTitle,
+  DialogTrigger,
+};
